Make JIT a runtime flag
Previously, the compile time flag will include generators inside init.c. Lack of generators means JIT is disabled. But now, with a runtime flag, we use the presence of code cache to indicate if JIT is enabled or not. If it is NULL, JIT is disabled, otherwise, JIT is enabled.

The main changes are in subgraph.h to make struct xnn_runtime take a pointer to code cache instead of holding the struct itself, so that it can be NULL.

The rest of the changes are in benchmarks. To avoid excessive duplication, FP16 and FP32 models take a new parameter `use_jit` to indicate if JIT should be enabled, then the E2E JIT tests will call a helper benchmark function that passes use_jit == true. We currently only have f32-gemm-e2e JIT benchmarks, f16-gemm-e2e will be added in a follow-up.

We don't have sparse (convolution-nchw) JIT support) so remove mentions of JIT entirely.

PiperOrigin-RevId: 518289830